CRM457-1467: Fix floating point rounding errors (#477)
## Description of change
We were explicitly casting some values as floats, and IntegerTimePeriod
multiplications were being quietly converted to floats even when
multiplied with decimals. This was leading to rounding errors.

[Link to relevant
ticket](https://dsdmoj.atlassian.net/browse/CRM457-1467)

+++ app/view_models/nsm/v1/work_item.rb
@@ -8,10 +8,10 @@ module Nsm
       adjustable_attribute :time_spent, :time_period
       attribute :completed_on, :date
 
-      attribute :pricing, :float
+      attribute :pricing, :decimal
       adjustable_attribute :uplift, :integer
       attribute :fee_earner, :string
-      attribute :vat_rate, :float
+      attribute :vat_rate, :decimal
       attribute :firm_office
 
       class << self
